First pass at test framework
