+++ .gitignore
@@ -1,3 +1,6 @@
+bin/
+.idea/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
