+++ requirements.txt
@@ -1,0 +1,3 @@
+html5lib
+pyRdfa3
+rdflib
